# [k8s] Use streaming json parser for listing nodes and pods: up to 2x speedup, 50% less memory for large clusters