`get_kubernetes_node_info` calls the k8s API to list all nodes and pods. On large kubernetes clusters, this can be expensive in many ways, as the size of the response body can get really big. On a mock API server setup using [kwok](https://kwok.sigs.k8s.io/), I mocked 500 nodes and 5000 pods. The response size is 4MB for list nodes, and 34MB for list pods. I would expect real clusters to be bigger than this, as real nodes and pods would contain a lot more labels, annotations, etc.

This large response size contributes to both increased memory usage, as we have to load everything to memory to deserialize it into python objects, and also increased CPU usage and latency, which is mainly caused by a long standing [issue](https://github.com/kubernetes-client/python/issues/1921) on the kubernetes python client.

This PR is an attempt to improve this by using a streaming json parser ([ijson](https://github.com/ICRAR/ijson)), which should limit memory usage (as the streaming json parser will only keep BUFFER_SIZE bytes in memory at any given time, which we can configure). It should also improve the latency, as we won't incur the cost of deserializing to the kubernetes client's python objects.

The main trade-off here is we need to maintain our own schema which is a subset of the k8s API schema. The good thing is we only need to do it for two resources, Node and Pod, both of which are stable APIs at this point. Plus we only care about a few fields.

Preliminary results of Peak RSS + latency on the mock API server:

| Test Case | Master | This PR |
|-----------|--------------|-------------|
| 500 nodes + 5000 pods | 474.98 MB (6.7s) | 271.50 MB (2s) |
| 500 nodes + 1000 pods | 265.23 MB (2.4s)  | 218.45 MB (1.1s) |
| 100 nodes + 1000 pods | 229.33 MB (1.3s) | 187.80 MB (0.5s) |
| 50 nodes + 500 pods | 199.41 MB (0.8s) | 176.38 MB (0.3s) |

Note: For the latency improvements, I feel like we will see similar numbers if we just use `json.loads()`, because the main bottleneck is on the k8s client deserializer.

Tested (run the relevant ones):

- [x] Code formatting: install pre-commit (auto-check on commit) or `bash format.sh`
- [x] Any manual or new tests for this PR (please specify below)
  - New smoke test that compares our custom parsing with the original deserializer from the kubernetes client package
- [x] All smoke tests: `/smoke-test` (CI) or `pytest tests/test_smoke.py` (local)
- [x] Relevant individual tests: `/smoke-test -k test_name` (CI) or `pytest tests/test_smoke.py::test_name` (local)
- [x] Backward compatibility: `/quicktest-core` (CI) or `pytest tests/smoke_tests/test_backward_compat.py` (local)

<!-- CI commands (/-prefixed) can only be triggered by repo members -->