@SeungjinYang 2025-10-03
/smoke-test --kubernetes