@kevinmingtarja 2025-10-04
/smoke-test -k test_kubernetes_get_nodes_and_pods --kubernetes