@kevinmingtarja 2025-10-04
/quicktest-core --kubernetes --base-branch master